1. fix: two separator lines 2. fix: padding of date input 3. fix: Image slider keep no margin space with navigator bar in mobile version 4. add: image slide auto-play 5. add: Python Flask Blueprint
