+++ templates/attraction.html
@@ -63,6 +63,7 @@
                 </div>
             </div>     
         </section>
+        <hr/>
         <section class="attraction-infos wrapper">
             <div class="description"></div>
             <div class="address">景點地址:</div>
